feat/kiwi-web: Add graphql query to create product.
Added @apollo/client and graphql packages to run graphql queries. This seems like the simplest way for now, while leaving potential for pagination, cache, and tests in the future.

The way the mutation is run will likely change, as I want to clean up how the data is passed, and potentially how the multipart form is handled.