refactor!(apteryx): Migrate graphql-modules to explicit gql types.
I can do it myself without the abstraction graphql-modules provides. It also allows me to use graphql-iso-date more cleanly, instead of hacking in bigint with floats to store unix timestamps instead of a readable date.

Still need to add mongoose integration with the resolvers, but all the queries kiwi-web will make are currently drafted in.

BREAKING CHANGE: Removing graphql-modules is a major change to the code structure, but shouldn't affect interaction WHEN COMPLETED. Currently there are no resolvers for mongoose, so nothing will resolve to the data you expect, just the placeholder information.